refactor galsim fits generation code to support different test applications
Previously we ran `generate_test_image.py`, which contained the outer program logic for FITS file
generation but imported all the test case definitions and galsim-specific code from
`test_case_definitions.py`. This change puts all the outer logic and all the galsim-specific code
into `generate_test_image.py`, which is not a library, not an executable
script. `test_case_definitions.py` has been moved to `galsim_benchmarks.py` and just defined the
test cases and calls into the shared library. Now we can add, e.g., `galsim_field.py`, which
generates a field of 500 sources into a different FITS file, and maybe other applications in the
future.
