--- benchmark/galsim/latest_fits_filename.txt
@@ -1,1 +1,0 @@
-galsim_test_images_7b461cae2a.fits
